Rename ome_zarr_multiscale.py to ome_zarr_helpers.py
