+++ webknossos/webknossos/dataset/_image_conversion/ome_zarr_helpers.py
