+++ webknossos/webknossos/dataset/_image_conversion/zarr_image_source.py
@@ -27,9 +27,9 @@ from ..defaults import (
 from ..errors import CorruptImageError
 from .image_source import ReadOptions, compute_channel_selection
 from .image_source_registry import register_chunked_image_source
-from .ome_zarr_multiscale import OmeChannelMetadata, resolve_ome_multiscale
-from .ome_zarr_multiscale import layer_split_label as _ome_layer_split_label
-from .ome_zarr_multiscale import (
+from .ome_zarr_helpers import OmeChannelMetadata, resolve_ome_multiscale
+from .ome_zarr_helpers import layer_split_label as _ome_layer_split_label
+from .ome_zarr_helpers import (
     suggested_view_configuration as _ome_suggested_view_configuration,
 )
 from .tensorstore_chunked_image_source import TensorStoreChunkedImageSource, guess_axes
